+++ packages/medusa-file-s3/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.9](https://github.com/medusajs/medusa/compare/medusa-file-s3@1.0.8...medusa-file-s3@1.0.9) (2021-12-08)
+
+**Note:** Version bump only for package medusa-file-s3
+
 ## [1.0.8](https://github.com/medusajs/medusa/compare/medusa-file-s3@1.0.7...medusa-file-s3@1.0.8) (2021-11-23)
 
 **Note:** Version bump only for package medusa-file-s3